+++ client/src/components/404/404page.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import App from "./App";
+import App from "../App";
 import {Link} from 'react-router-dom';
 
 function Page(): JSX.Element {
