+++ client/src/components/About/about.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
-import App from "./App";
-function Contact(): JSX.Element {
+import App from "../App";
+function About(): JSX.Element {
     return (
         <div>
             <App />
             <div className="container my-4 text-center">
-                <h2>Contact Us</h2>
-                <p>This ia a Contact page</p>
+                <h2>About Us</h2>
+                <p>This ia a About page</p>
             </div>
         </div>
     );
 }
-export default Contact;
+export default About;
