+++ client/src/components/Contact/contact.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
-import App from "./App";
-function Home(): JSX.Element {
+import App from "../App";
+function Contact(): JSX.Element {
     return (
         <div>
             <App />
             <div className="container my-4 text-center">
-                <h2>Home</h2>
-                <p>You have customer records</p>
+                <h2>Contact Us</h2>
+                <p>This ia a Contact page</p>
             </div>
         </div>
     );
 }
-export default Home;
+export default Contact;
