+++ client/src/components/Intro/intro.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import App from "./App";
+import App from "../App";
 import { Image } from "./Image";
 
 function Intro(): JSX.Element {
