+++ client/src/components/To-Do/To-Do.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import App from "./App";
+import App from "../App";
 import DateBar from "./DateBar";
 import TodoList from "./List";
 import Addactivity from "./Addactivity";
